Styles and winnings tracker
Styles in site.css weren't cooperating so most of them are inline.
Added the winning tracker and moved buttons to the nav bar in the header.

+++ BlackjackApp/Models/Player.cs
@@ -8,7 +8,7 @@ namespace BlackjackApp.Models
     public class Player
     {
         public Hand hand { get; set; }
-        public int winnings { get; set; }
+        public double winnings { get; set; }
 
         public Player()
         {
